New Alpine-Base Container for OS Base Installation Changed all other Dockerfiles to use this docker-alpine-base
Icinga2 and IcingaWeb2 was moved from testing repository to the community.
This was tricky for an dependancy deeplock at alpine
The Database-Schema was was moved to another directory

remove a hard coded container in the icinga-api Script

add clean mysql-configuration for better memory settings

include our Hackathon Presentation

the grafana container doesn't import dashboards anymore.
this is the job for add-host in the docker-cm-monitoring Container