now, we can overwrite the template for grafana
  (https://github.com/cm-xlabs/monitoring/issues/23)

basicly, the Grafana Templates will be created by a internal logic
the application-name is similar to the grafana-template name

but, a customer can change the application name and our logic will found no template.

now we can give every application an own template in cm-service.yaml:

  cae-live-1:
    description: CAE Live
    port: 42199
    cap_connection: true
    uapi_cache: true
    blob_cache: true
    application:
    - cae
    - caches
    - caches-ibm
    template: cae

at this point, the description and the template vars will be used to create an new dashboard.

the template must be located under 'docker-cm-monitoring/rootfs/usr/local/share/templates/grafana/services'
for the example above, the template must be named 'cae.json'